WORKING STATE
I created duplicate rows with the inner join around line 60. Joining on customer_id created a LOT of duplicate rows for each purchase. This is eliminated by joining on rental_id since rental_id is unique.

Was able to delete the DISTINCT clause from SELECT.

Fixed the count for the amount of rentals. Got rid of duplicate rows in summary table.